Update game account method.Implement delete image method

+++ AccountHub.Domain/Entities/AccountImageEntity.cs
@@ -5,12 +5,10 @@ namespace AccountHub.Domain.Entities;
 
 public class AccountImageEntity
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public required string ImageUrl { get; set; }
     
-    public int Order { get; set; } 
-    
-    
+    public int Order { get; set; }
     public long GameAccountId { get; set; }
     [JsonIgnore]
     public GameAccountEntity? GameAccount { get; set; }
